replace mysql with sqlite
the application no longer relies on mysql now so it's more self contained.
so the only external requirement is flowise flow (included).

this can also then be replaced with the langchain code (or a prompt specifially for outputting json in an output parser style output directly to openai or wherever).

TODO - add a "reset" within the UI to delete all the rows in the aiworld table.