some fixes
sometimes the api will copy the historical ability attacks which are like "attack:lucfer", so instead of putting the target, the response will be "attack:Lucifer" instead of "Lucifer".this fix keeps just an Entity name so it should cover most issues relating to this.

added more bots for better testing.