removed custom LLMs . Changed Abilities to Actions
Flowise is just so easy to maintain than manually doing all the output formats / langchain type stuff.
So no you will need to use flowise or use the old scripts.

Also changed abilities to actions. This will allow use to reduce the output fields from the LLM, for example, pickup item, or other types of actions that we can to add later.