move is now an action
this may have been a "bad" move.
bots can now only perform one action at a time - e.g. move, attack or heal.

this stops them from running off, which is probably a bad thing. We will see. We will also add pickup item as an action as well.

currently with this change, the bots have a hard time picking up items. We'll see how it works out when we add pickup as an action.